# openapi.yaml
openapi: 3.0.0
info:
  title: Treza Platform API
  description: |
    Comprehensive API for managing secure enclaves, GitHub integrations, and platform resources.
    
    **New in v2.1.0:**
    - 🛡️ Cryptographic attestation documents and verification
    - 📋 Developer integration tools and API endpoints
    - 🏦 Use case templates for Financial, Healthcare, and Enterprise
    - 🔐 PCR measurements and security validation
    - 📊 Business impact and compliance tracking
    
    **Previous in v2.0.0:**
    - 🔄 Enclave lifecycle management (pause, resume, terminate)
    - 📜 Comprehensive logs access (application, ECS, Step Functions, Lambda, errors)
    - 🐳 Docker Hub integration (search images, get tags)
    - 📊 Enhanced status tracking with real-time updates
    - 🎯 Individual enclave management endpoints
  version: 2.1.0
  contact:
    name: Treza Support
    url: https://trezalabs.com
    email: support@trezalabs.com
servers:
  - url: https://app.trezalabs.com
    description: Production server
  - url: http://localhost:3000
    description: Development server
paths:
...
  /api/enclaves/{id}/logs:
    get:
      summary: Get enclave logs
      description: Retrieve logs for a specific enclave from various sources
      tags:
        - Enclaves
        - Logs
      parameters:
        - name: id
          in: path
          required: true
          description: Enclave ID
          schema:
            type: string
            example: "enc_123456"
        - name: type
          in: query
          required: false
          description: Type of logs to retrieve
          schema:
            type: string
            enum: [all, ecs, stepfunctions, lambda, application, errors]
            default: all
            example: "application"
        - name: limit
          in: query
          required: false
          description: Maximum number of log entries to return
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
            example: 50
      responses:
        '200':
          description: Logs retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogsResponse'
        '404':
          description: Enclave not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Error message"
        details:
          type: array
          items:
            type: string
          description: "Additional error details (optional)"
    LogEntry:
      type: object
      properties:
        timestamp:
          type: integer
          format: int64
          example: 1640995200000
          description: "Unix timestamp in milliseconds"
        message:
          type: string
          example: "Application started successfully"
        source:
          type: string
          enum: [ecs, stepfunctions, lambda, application, dynamodb]
          example: "application"
        stream:
          type: string
          example: "ecs/treza-dev-terraform-runner/12345"
          description: "Log stream name (optional)"
        type:
          type: string
          enum: [application, stdout, stderr, info, error]
          example: "stdout"
          description: "Log type (optional)"
        logGroup:
          type: string
          example: "/aws/nitro-enclave/enc_123456/application"
          description: "CloudWatch log group (optional)"
        function:
          type: string
          example: "treza-dev-validation"
          description: "Lambda function name for lambda logs (optional)"
        execution:
          type: string
          example: "deployment-enc_123456"
          description: "Step Function execution name (optional)"
        stateMachine:
          type: string
          enum: [deployment, cleanup]
          example: "deployment"
          description: "Step Function state machine type (optional)"
    LogsResponse:
      type: object
      properties:
        enclave_id:
          type: string
          example: "enc_123456"
        enclave_name:
          type: string
          example: "Trading Bot Enclave"
        enclave_status:
          type: string
          example: "DEPLOYED"
        logs:
          type: object
          properties:
            ecs:
              type: array
              items:
                $ref: '#/components/schemas/LogEntry'
              description: "ECS deployment logs (optional)"
            stepfunctions:
              type: array
              items:
                $ref: '#/components/schemas/LogEntry'
              description: "Step Functions workflow logs (optional)"
            lambda:
              type: array
              items:
                $ref: '#/components/schemas/LogEntry'
              description: "Lambda function logs (optional)"
            application:
              type: array
              items:
                $ref: '#/components/schemas/LogEntry'
              description: "Application logs from the enclave (optional)"
            errors:
              type: array
              items:
                $ref: '#/components/schemas/LogEntry'
              description: "Error logs from all sources (optional)"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    
    WalletAuth:
      type: apiKey
      in: query
      name: wallet
      description: Wallet address for authentication
security:
  - BearerAuth: []
  - WalletAuth: []
...